Add support for arbitrary float day deadlines
STs have requested the ability to set deadlines at half-hour increments.
This implementation is simpler than requiring the length to be a whole
or half integer, and imposes no constraints on usability or additional
complexity.

+++ lib/cogs/game_progression.py
@@ -215,7 +215,7 @@ class GameProgression(commands.Cog, name="Game Progression"):
     @checks.is_game()
     @checks.is_storyteller()
     @checks.is_dm()
-    async def setdeadline(self, ctx: Context, length: int):
+    async def setdeadline(self, ctx: Context, length: float):
         """Set a deadline for the current day and open nominations.
 
         length: The number of hours for the deadline to last.
